# Nadam optimizer and test for it added