@the-moliver 2016-05-20
Nice work! I would suggest you expose the three constants you have hardcoded in these lines so they can be modified by the user:
```python
schedule_decay = 0.004  # Exactly given in [1] and [2]
momentum_cache_t = self.beta_1 * (1. - 0.5 * (K.pow(0.96, t * schedule_decay)))
momentum_cache_t_1 = self.beta_1 * (1. - 0.5 * (K.pow(0.96, (t + 1) * schedule_decay)))
```