@iikulikov 2016-05-21
@the-moliver thanks! I also thought about that and when I did more experiments some time ago not with keras, I tried to jiggle this values without any improvements even more, usually I got worse performance.

This constants came from [here](http://cs229.stanford.edu/proj2015/054_report.pdf), page 4, equation 5, authors said that this constants came from earlier Nesterov papers (one of them is from 1983) and he motivated these values also by some deeper research.

If @fchollet ask to expose this variables too, then I will do that. But I think the best is to give meaningful comment in documentation about these values. (I actually will do that in doc-string I guess).