@fchollet 2016-06-11
After external feedback, it seems like this is pretty cool and we'll merge it. Thanks!